By now, playlist function is still in progress.
But, the scrolling or swipe in viewpager is solved.
Previously, there is a problem in swiping list item. Conflicts exist between swiping list item and scroll page.
To solve, I adjust the sensitive area of swiping a list item.
So that when user scroll or swipe from the right part of the monitor, page is scrolling.
And if user swipe near middle or left part, list item was swiped.

Added "enable left swipe mode" for swiping a list item.